polish paths of shared functions and config files
