+++ tests/cypress/common/tests.js
@@ -6,9 +6,10 @@ import {
   verifyPolicyInPolicyDetailsTemplates, verifyPlacementRuleInPolicyDetails, verifyPlacementBindingInPolicyDetails,
   verifyViolationsInPolicyStatusClusters, verifyViolationsInPolicyStatusTemplates,
   getViolationsPerPolicy, getViolationsCounter, verifyPolicyDetailsInCluster, verifyPolicyTemplatesInCluster,
-  verifyPolicyViolationDetailsInCluster, verifyPolicyViolationDetailsInHistory
-} from '../../views/policy'
-import { getConfigObject } from '../../config'
+  verifyPolicyViolationDetailsInCluster, verifyPolicyViolationDetailsInHistory,
+  createPolicyFromYAML
+} from './views'
+import { getConfigObject } from '../config'
 
 
 export const test_genericPolicyGovernance = (confFilePolicy, confFileViolationsInform, confFileViolationsEnforce=null, confFileClusters='clusters.yaml', filteredClusterList=null) => {
@@ -244,5 +245,38 @@ export const test_genericPolicyGovernance = (confFilePolicy, confFileViolationsI
       verifyPolicyNotInListing(policyName)
     })
   }
+
+}
+
+
+export const test_applyPolicyYAML = (confFilePolicy, substitutionRules=null) => {
+
+  if (substitutionRules === null) {
+    substitutionRules = getDefaultSubstitutionRules()
+  }
+  const rawPolicyYAML = getConfigObject(confFilePolicy, 'raw', substitutionRules)
+  const policyName = rawPolicyYAML.replace(/\r?\n|\r/g, ' ').replace(/^.*?name:\s*/m, '').replace(/\s.*/m,'')
+
+  it('Create the clean up policy using the YAML', () => {
+    cy.visit('/multicloud/policies/create')
+    cy.log(rawPolicyYAML)
+    createPolicyFromYAML(rawPolicyYAML, true)
+  })
+
+  it(`Check that policy ${policyName} is present in the policy listing`, () => {
+    verifyPolicyInListing(policyName, {})
+  })
+
+  it(`Wait for ${policyName} status to become available`, () => {
+    cy.waitForPolicyStatus(policyName, '0/')
+  })
+
+  it(`Delete policy ${policyName}`, () => {
+    actionPolicyActionInListing(policyName, 'Remove')
+  })
+
+  it(`Verify that policy ${policyName} is not present in the policy listing`, () => {
+    verifyPolicyNotInListing(policyName)
+  })
 
 }
