+++ tests/cypress/config/demos/policy-demo/demo-policy-config.yaml
