+++ tests/cypress/config/demos/policy-demo/demo-policy-placement-rule.yaml
