+++ tests/cypress/tests/IamPolicy_governance.spec.js
@@ -1,10 +1,9 @@
 /* Copyright (c) 2020 Red Hat, Inc. */
 /// <reference types="cypress" />
-import { test_genericPolicyGovernance } from './common/generic_policies_governance'
-import { cleanup_usingPolicyYAML } from './common/generic_policy_cleanup'
+import { test_genericPolicyGovernance, test_applyPolicyYAML } from '../common/tests'
 
 describe('GRC UI: [P1][Sev1][policy-grc] IamPolicy governance - setup', () => {
-  cleanup_usingPolicyYAML('IamPolicy_governance/policy-clusterrolebinding-setup_raw.yaml')
+  test_applyPolicyYAML('IamPolicy_governance/policy-clusterrolebinding-setup_raw.yaml')
 })
 
 describe('RHACM4K-1719 - GRC UI: [P1][Sev1][policy-grc] IamPolicy governance', () => {
@@ -12,5 +11,5 @@ describe('RHACM4K-1719 - GRC UI: [P1][Sev1][policy-grc] IamPolicy governance', (
 })
 
 describe('GRC UI: [P1][Sev1][policy-grc] IamPolicy governance - cleanup', () => {
-  cleanup_usingPolicyYAML('IamPolicy_governance/policy-clusterrolebinding-cleanup_raw.yaml')
+  test_applyPolicyYAML('IamPolicy_governance/policy-clusterrolebinding-cleanup_raw.yaml')
 })
