+++ tests/cypress/tests/ImageManifest_governance.spec.js
@@ -1,8 +1,7 @@
 /* Copyright (c) 2020 Red Hat, Inc. */
 /// <reference types="cypress" />
-import { test_genericPolicyGovernance } from './common/generic_policies_governance.js'
+import { test_genericPolicyGovernance, test_applyPolicyYAML } from '../common/tests.js'
 import { getClusterListByVendor } from '../config'
-import { cleanup_usingPolicyYAML } from './common/generic_policy_cleanup'
 
 const getList = getClusterListByVendor('OpenShift')
 
@@ -16,6 +15,6 @@ if (getList.length>0) {
   test_genericPolicyGovernance('ImageManifest_governance/policy-config.yaml', 'ImageManifest_governance/violations-inform.yaml', 'ImageManifest_governance/violations-enforce.yaml', 'clusters.yaml', getList)
 })
   describe('GRC UI: [P1][Sev1][policy-grc] ImageManifest governance - cleanup', () => {
-    cleanup_usingPolicyYAML('ImageManifest_governance/ImageManifest_specification_cleanup_policy_raw.yaml')
+    test_applyPolicyYAML('ImageManifest_governance/ImageManifest_specification_cleanup_policy_raw.yaml')
   })
 }
