+++ tests/cypress/tests/LimitRange_governance.spec.js
@@ -1,10 +1,9 @@
 /* Copyright (c) 2020 Red Hat, Inc. */
 /// <reference types="cypress" />
-import { test_genericPolicyGovernance } from './common/generic_policies_governance'
-import { cleanup_usingPolicyYAML } from './common/generic_policy_cleanup'
+import { test_genericPolicyGovernance, test_applyPolicyYAML } from '../common/tests'
 
 describe('GRC UI: [P1][Sev1][policy-grc] LimitRange policy governance - setup', () => {
-    cleanup_usingPolicyYAML('LimitRange_governance/LimitRange_specification_setup_policy_raw.yaml')
+    test_applyPolicyYAML('LimitRange_governance/LimitRange_specification_setup_policy_raw.yaml')
 })
 
 describe('RHACM4K-1726 - GRC UI: [P1][Sev1][policy-grc] LimitRange policy governance', () => {
@@ -12,5 +11,5 @@ describe('RHACM4K-1726 - GRC UI: [P1][Sev1][policy-grc] LimitRange policy govern
 })
 //setup & clean up yaml files are identical
 describe('GRC UI: [P1][Sev1][policy-grc] LimitRange policy governance - clean up', () => {
-    cleanup_usingPolicyYAML('LimitRange_governance/LimitRange_specification_setup_policy_raw.yaml')
+    test_applyPolicyYAML('LimitRange_governance/LimitRange_specification_setup_policy_raw.yaml')
 })
