+++ tests/cypress/tests/create_policy_multiselect_sanity.spec.js
@@ -1,7 +1,7 @@
 /* Copyright (c) 2020 Red Hat, Inc. */
 /// <reference types="cypress" />
 
-import { checkItems, verifyItemsChecked } from '../views/common'
+import { checkItems, verifyItemsChecked } from '../common/views'
 
 const specificationsList = ['CertificatePolicy', 'EtcdEncryption']
 const selectSpecsQuery = '.bx--multi-select[aria-label="specs"]'
