+++ tests/cypress/tests/demos/policy-demo.spec.js
@@ -5,7 +5,7 @@ import {
   actionPolicyActionInListing, getDefaultSubstitutionRules,
   verifyPlacementRuleInPolicyDetails, verifyPlacementBindingInPolicyDetails,
   verifyPolicyInPolicyDetails, verifyPolicyInPolicyDetailsTemplates
-} from '../../views/policy'
+} from '../../common/views'
 import { getUniqueResourceName } from '../../scripts/utils'
 import { getConfigObject } from '../../config'
 
@@ -14,11 +14,11 @@ describe('Testing policy named demo-policy in demo.yaml file', () => {
     const uPolicyName = getUniqueResourceName(policyName)
     // demo-policy-raw.yaml is used for creating the policy "demo-policy"
     // demo-policy-raw.yaml is raw policy yaml and need be to get as raw data
-    const policyYAML = getConfigObject('sample/demo-policy-raw.yaml', 'raw', getDefaultSubstitutionRules({policyname:uPolicyName}))
+    const policyYAML = getConfigObject('demos/policy-demo/demo-policy-raw.yaml', 'raw', getDefaultSubstitutionRules({policyname:uPolicyName}))
     // demo-policy-config.yaml is used for validating the policy "demo-policy"
     // demo-policy-config.yaml isn't raw policy yaml but config yaml and need be converted to a dictionary
-    const { policyConfig } = getConfigObject('sample/demo-policy-config.yaml')
-    const confClusterViolations = getConfigObject('sample/violations.yaml', 'yaml', getDefaultSubstitutionRules({policyname:policyName}))
+    const { policyConfig } = getConfigObject('demos/policy-demo/demo-policy-config.yaml')
+    const confClusterViolations = getConfigObject('demos/policy-demo/violations.yaml', 'yaml', getDefaultSubstitutionRules({policyname:policyName}))
 
     it (`Can create new policy ${uPolicyName} from YAML editor`, () => {
       cy.FromGRCToCreatePolicyPage()
